requirements.txt
Flask==2.1.1
pymorphy2==0.9.1
requests~=2.27.1
python-dotenv~=0.20.0
redis~=4.2.2
